# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Openclaw Agent OS API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /vps/config:
    get:
      operationId: getVpsConfig
      tags: [vps]
      summary: Get VPS connection configuration
      responses:
        "200":
          description: VPS config (credentials redacted)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VpsConfig"
        "404":
          description: Not configured
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    put:
      operationId: saveVpsConfig
      tags: [vps]
      summary: Save VPS SSH connection config
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SaveVpsConfigBody"
      responses:
        "200":
          description: Saved config
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VpsConfig"
...
  /vps/processes/{pid}/kill:
    post:
      operationId: killVpsProcess
      tags: [vps]
      summary: Kill a process by PID
      parameters:
        - name: pid
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VpsCommandResult"
...
  /vps/services/{name}/{action}:
    post:
      operationId: controlVpsService
      tags: [vps]
      summary: Start, stop, or restart a service
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
        - name: action
          in: path
          required: true
          schema:
            type: string
            enum: [start, stop, restart, status]
      responses:
        "200":
          description: Result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VpsCommandResult"
  /vps/files:
    get:
      operationId: listVpsFiles
      tags: [vps]
      summary: List files in a VPS directory
      parameters:
        - name: path
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: File list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/VpsFileEntry"
  /vps/files/read:
    get:
      operationId: readVpsFile
      tags: [vps]
      summary: Read a file from the VPS
      parameters:
        - name: path
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: File content
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VpsFileContent"
  /vps/files/write:
    post:
      operationId: writeVpsFile
      tags: [vps]
      summary: Write a file to the VPS
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WriteVpsFileBody"
      responses:
        "200":
          description: Write result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VpsCommandResult"
  /vps/files/delete:
    delete:
      operationId: deleteVpsFile
      tags: [vps]
      summary: Delete a file or directory from the VPS
      parameters:
        - name: path
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Delete result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VpsCommandResult"
  /vps/exec:
    post:
      operationId: execVpsCommand
      tags: [vps]
      summary: Execute a shell command on the VPS
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExecVpsCommandBody"
      responses:
        "200":
          description: Command output
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VpsCommandResult"
  /vps/logs:
    get:
      operationId: tailVpsLog
      tags: [vps]
      summary: Tail a log file on the VPS (SSE stream)
      parameters:
        - name: path
          in: query
          required: true
          schema:
            type: string
        - name: lines
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: SSE stream of log lines
          content:
            text/event-stream: {}

  # ── Settings ─────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    ApiError:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    # ── Agent schemas ──────────────────────────────────────────────────────────
...
    VpsConfig:
      type: object
      properties:
        id:
          type: integer
        label:
          type: string
        host:
          type: string
        port:
          type: integer
        username:
          type: string
        authType:
          type: string
          enum: [password, key]
        hasCredentials:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - label
        - host
        - port
        - username
        - authType
        - hasCredentials
        - createdAt
    SaveVpsConfigBody:
      type: object
      properties:
        label:
          type: string
        host:
          type: string
        port:
          type: integer
        username:
          type: string
        authType:
          type: string
          enum: [password, key]
        password:
          type: string
          nullable: true
        privateKey:
          type: string
          nullable: true
      required:
        - label
        - host
        - port
        - username
        - authType
...
    VpsFileEntry:
      type: object
      properties:
        name:
          type: string
        path:
          type: string
        type:
          type: string
          enum: [file, directory, symlink]
        size:
          type: integer
        modifiedAt:
          type: string
      required:
        - name
        - path
        - type
        - size
        - modifiedAt
    VpsFileContent:
      type: object
      properties:
        path:
          type: string
        content:
          type: string
        encoding:
          type: string
      required:
        - path
        - content
        - encoding
    WriteVpsFileBody:
      type: object
      properties:
        path:
          type: string
        content:
          type: string
      required:
        - path
        - content
    ExecVpsCommandBody:
      type: object
      properties:
        command:
          type: string
        timeout:
          type: integer
          nullable: true
      required:
        - command
    VpsCommandResult:
      type: object
      properties:
        success:
          type: boolean
        stdout:
          type: string
        stderr:
          type: string
        exitCode:
          type: integer
          nullable: true
      required:
        - success
        - stdout
        - stderr

    # ── Settings schemas ───────────────────────────────────────────────────────
...